# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Flask Microservices API
  version: 1.0.0
  description: Users and orders microservices API built with Flask.
servers:
  - url: http://127.0.0.1:5001
    description: user-service local server
  - url: http://127.0.0.1:5002
    description: order-service local server
tags:
  - name: Users
  - name: Orders
  - name: System
paths:
  /health:
    get:
      tags: [System]
      summary: Service health check
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
  /users:
    get:
      tags: [Users]
      summary: List users
      servers:
        - url: http://127.0.0.1:5001
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
    post:
      tags: [Users]
      summary: Create user
      servers:
        - url: http://127.0.0.1:5001
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
  /users/{user_id}:
    get:
      tags: [Users]
      summary: Get user by ID
      servers:
        - url: http://127.0.0.1:5001
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          $ref: '#/components/responses/NotFound'
  /ready:
    get:
      tags: [System]
      summary: order-service readiness check
      servers:
        - url: http://127.0.0.1:5002
      responses:
        '200':
          description: Service and dependency are reachable
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /config:
    get:
      tags: [System]
      summary: order-service runtime configuration
      servers:
        - url: http://127.0.0.1:5002
      responses:
        '200':
          description: Runtime configuration
  /orders:
    get:
      tags: [Orders]
      summary: List orders
      servers:
        - url: http://127.0.0.1:5002
      responses:
        '200':
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
    post:
      tags: [Orders]
      summary: Create order
      servers:
        - url: http://127.0.0.1:5002
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCreate'
      responses:
        '201':
          description: Order created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          $ref: '#/components/responses/BadRequest'
        '502':
          $ref: '#/components/responses/BadGateway'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /orders/{order_id}:
    get:
      tags: [Orders]
      summary: Get order by ID
      servers:
        - url: http://127.0.0.1:5002
      parameters:
        - in: path
          name: order_id
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: Order found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          $ref: '#/components/responses/NotFound'
components:
  schemas:
    HealthResponse:
      type: object
      properties:
        status:
          type: string
          example: ok
        service:
          type: string
          example: user-service
    UserCreate:
      type: object
      required: [name, email]
      properties:
        name:
          type: string
          example: Alice
        email:
          type: string
          format: email
          example: alice@example.com
    User:
      type: object
      required: [id, name, email]
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: Alice
        email:
          type: string
          format: email
          example: alice@example.com
    OrderCreate:
      type: object
      required: [user_id, item, price]
      properties:
        user_id:
          type: integer
          minimum: 1
          example: 1
        item:
          type: string
          example: Book
        price:
          type: number
          format: float
          minimum: 0
          example: 10.5
    Order:
      type: object
      required: [id, user_id, item, price, user_name, user_email, created_at]
      properties:
        id:
          type: integer
          example: 1
        user_id:
          type: integer
          example: 1
        item:
          type: string
          example: Book
        price:
          type: number
          example: 10.5
        user_name:
          type: string
          example: Alice
        user_email:
          type: string
          example: alice@example.com
        created_at:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: BAD_REQUEST
            message:
              type: string
              example: Field 'name' must be a non-empty string
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    BadGateway:
      description: Invalid dependency response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ServiceUnavailable:
      description: Dependency unavailable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'